# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VoiceConfigOut": {
        "properties": {
          "available": {
            "default": true,
            "title": "Available",
            "type": "boolean"
          },
          "backend": {
            "enum": [
              "offline",
              "cloud"
            ],
            "title": "Backend",
            "type": "string"
          },
          "boost_terms": {
            "title": "Boost Terms",
            "type": "integer"
          },
          "lasa_pairs": {
            "title": "Lasa Pairs",
            "type": "integer"
          },
          "lexicon_coverage": {
            "title": "Lexicon Coverage",
            "type": "number"
          },
          "lexicon_entries": {
            "title": "Lexicon Entries",
            "type": "integer"
          },
          "sample_rate": {
            "default": 16000,
            "title": "Sample Rate",
            "type": "integer"
          },
          "stt_model": {
            "title": "Stt Model",
            "type": "string"
          },
          "stt_provider": {
            "title": "Stt Provider",
            "type": "string"
          },
          "tts_model": {
            "title": "Tts Model",
            "type": "string"
          },
          "tts_provider": {
            "title": "Tts Provider",
            "type": "string"
          },
          "tts_quality": {
            "enum": [
              "flash",
              "multilingual"
            ],
            "title": "Tts Quality",
            "type": "string"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "ws_path": {
            "default": "/api/v1/voice/ws",
            "title": "Ws Path",
            "type": "string"
          }
        },
        "required": [
          "backend",
          "stt_provider",
          "stt_model",
          "tts_provider",
          "tts_model",
          "tts_quality",
          "boost_terms",
          "lexicon_entries",
          "lexicon_coverage",
          "lasa_pairs"
        ],
        "title": "VoiceConfigOut",
        "type": "object"
      },
      "VoiceSettingsIn": {
        "properties": {
          "tts_quality": {
            "enum": [
              "flash",
              "multilingual"
            ],
            "title": "Tts Quality",
            "type": "string"
          }
        },
        "required": [
          "tts_quality"
        ],
        "title": "VoiceSettingsIn",
        "type": "object"
      },
      "VoiceSettingsOut": {
        "properties": {
          "tts_quality": {
            "enum": [
              "flash",
              "multilingual"
            ],
            "title": "Tts Quality",
            "type": "string"
          }
        },
        "required": [
          "tts_quality"
        ],
        "title": "VoiceSettingsOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Evidence-grounded clinical question answering over public medical literature.",
    "title": "ClinicalContext AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/voice/analytics": {
      "get": {
        "description": "Aggregate latency per leg, speculation economics, confirmations,\nbarge-ins, masks, and the cost of interactivity (11G.5, 11D.4).",
        "operationId": "voice_analytics_api_v1_voice_analytics_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceAnalyticsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Voice Analytics",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/config": {
      "get": {
        "description": "What the client is talking to: backend, models, vocabulary sizes.",
        "operationId": "voice_config_api_v1_voice_config_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Voice Config",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/settings": {
      "get": {
        "operationId": "get_settings__api_v1_voice_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings ",
        "tags": [
          "voice"
        ]
      },
      "patch": {
        "description": "Flash (latency) vs Multilingual (prosody): an owner-level org setting.",
        "operationId": "set_settings_api_v1_voice_settings_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Settings",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/speak": {
      "post": {
        "description": "Text \u2192 speech, as a WAV clip the browser plays: the chat reads its\nanswers aloud a few sentences at a time with this. Deepgram Aura-2 on\nthe deployed server, the operating system's voice locally.",
        "operationId": "speak_api_v1_voice_speak_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Speak",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/transcribe": {
      "post": {
        "description": "A spoken question from the chat \u2192 its text: voice typing, and each\nturn of a voice conversation. Deepgram's medical model on the deployed\nserver, faster-whisper locally; a server with neither says so.",
        "operationId": "transcribe_api_v1_voice_transcribe_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Transcribe Api V1 Voice Transcribe Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transcribe",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/turns": {
      "get": {
        "description": "Per-turn waterfalls (the session view's debug surface).",
        "operationId": "list_turns_api_v1_voice_turns_get",
        "parameters": [
          {
            "in": "query",
            "name": "query_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query Session Id"
            }
          },
          {
            "in": "query",
            "name": "voice_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Voice Session Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTurnList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Turns",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/voices": {
      "get": {
        "description": "The read-aloud voices a person can choose in Settings. Only the\nDeepgram speech service offers a choice; elsewhere the list is shown but\nthe server's own voice speaks, and ``selectable`` says so.",
        "operationId": "list_voices_api_v1_voice_voices_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoicesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Voices",
        "tags": [
          "voice"
        ]
      }
    },
...
  }
}